cp_backend/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1.1
django-cors-headers==3.5.0
django-extensions==3.0.6
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
idna==2.10
psycopg2==2.8.5
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
stripe==2.51.0
urllib3==1.25.10
